+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 1.0.0-beta.2 (Unreleased)
 
+### Breaking Changes
+
+- Moved `Azure::Identity::Version`, defined in `azure/identity/version.hpp` to `Azure::Identity::Details` namespace in `azure/identity/version.hpp` header.
+
 ### Other changes and Improvements
 
 - Add high-level and simplified identity.hpp file for simpler include experience for customers.